requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
coverage==7.6.12
cryptography==44.0.1
distlib==0.3.9
Django==5.1.6
django-picklefield==3.2
django-q2==1.7.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
filelock==3.17.0
identify==2.6.7
idna==3.10
iniconfig==2.0.0
msal==1.31.1
nodeenv==1.9.1
packaging==24.2
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
pycparser==2.22
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.10.0
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
virtualenv==20.29.2
